refactor(scratchpad): fold single-consumer modules into their sole callers
Several scratchpad modules held one class/function group used by exactly one
other module. Each now lives next to its only caller instead of behind a
separate import:

* passes.py's ScratchpadOptimizationPass -> allocator.py (coarse_tiling.py
  updates its import accordingly).
* op_features.py split by consumer: with_residency -> cost_objective.py,
  features_for_division/features_for_menu/_work_slices -> sa_cooptimizer.py
  (dropping the now-redundant lazy import in _bundle_objective).
  tests/inductor/test_op_features.py keeps testing both together against one
  shared fixture, since division features and residency are the two halves of
  the same candidate-scoring story.

Also completes an already in-progress cooptimization_scorer.py -> utils.py
fold: sa_cooptimizer.py's _precompute_spill_costs still called the deleted
module (scorer.hbm_bytes_per_us()), which NameError'd every solve and failed
34 tests.
