--- torch_spyre/_inductor/scratchpad/passes.py
@@ -1,46 +1,0 @@
-# Copyright 2026 The Torch-Spyre Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from abc import ABC, abstractmethod
-
-from torch._inductor.graph import GraphLowering
-from torch._inductor.ops_handler import WrapperHandler
-
-
-class ScratchpadOptimizationPass(ABC):
-    """
-    Abstract class for optimization passes which are implemented to improve
-    a graph's overall scratchpad memory utilization and/or memory latency.
-    """
-
-    @abstractmethod
-    def apply_pass(self, graph: GraphLowering):
-        """
-        Accepts a candidate graph to be optimized and evaluated for scratchpad memory allocation.
-        `graph` will be mutated according in an implementation defined way. The order and
-        number of nodes in the graph may change as a result of an optimization pass.
-
-        Args:
-            graph (GraphLowering): The graph to be optimized for scratchpad memory allocation
-        """
-        pass
-
-
-class _NameSwapHandler(WrapperHandler):
-    def __init__(self, inner, name_map: dict[str, str]):
-        super().__init__(inner)
-        self._name_map = name_map
-
-    def load(self, name, index):
-        return super().load(self._name_map.get(name, name), index)
